matiushin_medvedev/backend/requirements.txt
flask==1.1.2
psycopg2-binary==2.8.6
flask_api==2.0
flask_sqlalchemy==2.4.4
marshmallow==2.21.0
